refactor: :card_file_box: se ordenaron los archivos asociados a el sidebar, segun la estructura de arquitectura clean
se ordenaron los archivos asociados a el sidebar, segun la estructura de arquitectura clean
